+++ shell/commands/arp/ether.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+arp -n -H ether