+++ shell/commands/ethtool/base.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+set -e
+
+IFLIST=( $(ls /sys/class/net/) )
+
+for iface in ${IFLIST[@]}
+do
+	sudo ethtool ${iface}
+done